initial commit
In this file, we are simulating a convolutional neural network which is a topic in deep learning. We are using this convolutional neural network to classify images. The dataset we are using is “Brain MRI Images for Brain Tumor Detection”. The simulation is done to understand the working of deep neural networks in classification of images. In the result we have shown the performance of the neural network through many metrics like “accuracy”, “precision”, “recall” etc. The result also discusses the significance of convolutional neural networks in the pool of artificial neural networks in classification of images.